Added possibility to load images and smileys instead of dots
